Introduced new variable for server start time
Added a new variable in DockerSessionFactory that can be used as a maximum to wait for the server to start. Made it adjustable via a DockerFlag and set the default to 60 seconds, which was previously the hardcoded value in that field.

Also added a line to Dockerfile, because dev container was not working as is.